requirements.txt
aioredis==1.3.1
asgiref==2.3.2
async-timeout==2.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
cffi==1.13.2
channels==2.3.1
channels-redis==2.3.2
constantly==15.1.0
cryptography==2.8
daphne==2.4.0
Django==2.2.9
django-filter==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
djoser==2.0.3
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Markdown==3.1.1
msgpack==0.5.6
Pillow==6.2.1
pkg-resources==0.0.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2019.3
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
zope.interface==4.7.1
